backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.22
google-generativeai>=0.5.2
langchain>=0.1.0
langchain-community
langchain-google-genai
langchain-chroma
pyyaml==6.0.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0

